feat(api-worker-split): OOM mitigation via process role split, lazy Puppeteer, and telemetry gating
## Summary
Implements RPAS-CM governed feature 'api-worker-split' to mitigate
production OOM errors on 512MB Railway instances (REF: 2026-06-15 spec).

## Changes

### Process Role Split (REQ-001 to REQ-003)
- server/src/utils/processRole.ts: New utility - getProcessRole(),
  shouldRunWorkers(), shouldRunWebServer() driven by ADPA_PROCESS_ROLE env
- server/src/startup/serverBootstrap.ts: Gates Express boot on shouldRunWebServer()
- server/src/startup/dependencies/workers.ts: Gates queue consumer init on shouldRunWorkers()
- server/src/startup/dependencyGraph.ts: Role-aware dependency graph
- server/src/worker.ts: Lightweight worker-only bootstrap entrypoint
- server/package.json: Added 'worker' start script

### Queue Refactor (REQ-002)
- server/src/services/queue/queueClient.ts: Connection + client ops (publish, ack, nack)
- server/src/services/queue/registerWorkers.ts: Consumer bindings
- server/src/services/queueService.ts: Delegates to queueClient; lazy worker registration

### Lazy Puppeteer (REQ-004)
- server/src/services/pdfService.ts: Dynamic await import('puppeteer') inside
  getBrowser(); added cleanup() method; type-only import at top-level

### Telemetry Gating (REQ-005 & REQ-006)
- server/src/services/unifiedAIService.ts: Returns Langfuse trace/generation IDs
- server/src/services/documentGenerationService.ts: recordLLMPromptSnapshot()
  strips prompt/response blobs unless LLM_INSIGHTS_STORE_BLOBS=true
- server/src/routes/health.ts: Backend health route for dev banner

### Jest / ESM Fix
- server/jest.config.js, jest.config.unit.js: Platform-independent
  transformIgnorePatterns for uuid (node_modules[\\/]); added .js transform rule

### Governed Feature Registration
- server/governed-features.manifest.json: Registered 'api-worker-split' packet
- .agents/skills/adpa-api-worker-split/SKILL.md: Agent skill documentation
- docs/superpowers/specs/2026-06-15-api-worker-split-design.md: Design spec

## Test Results
npm run test:features -- api-worker-split
10/10 tests passing (REQ-001 through REQ-006 contracts GREEN)

## Environment Variables
- ADPA_PROCESS_ROLE: 'api' | 'worker' | 'all' (default: 'all' — no breaking change)
- LLM_INSIGHTS_STORE_BLOBS: 'true' to persist prompt/response blobs to Postgres

RPAS-CM: AEV-validated, governed-features manifest registered
