+++ .gitignore
@@ -205,3 +205,7 @@ scripts/*.log
 
 # Local logs
 logs/
+
+
+#Ignore vscode AI rules
+.github\instructions\codacy.instructions.md
