+++ server/src/routes/__tests__/health_logic.test.ts
@@ -2,15 +2,26 @@ import express from 'express';
 import request from 'supertest';
 import healthRoutes, { initializeDependencyHealthTracking, updateDependencyHealth } from '../health';
 
+const mockSafeQuery = jest.fn();
+
+jest.mock('../../services/jobs/dbGuards', () => ({
+  safeQuery: (...args: unknown[]) => mockSafeQuery(...args),
+}));
+
 describe('Health Endpoints Verification', () => {
   let app: express.Express;
 
   beforeAll(() => {
     app = express();
-    app.use('/', healthRoutes);
+    app.use('/health', healthRoutes);
     
     // Initialize with standard dependencies
     initializeDependencyHealthTracking(['Database', 'Redis', 'Neo4j', 'Pinecone']);
+  });
+
+  beforeEach(() => {
+    mockSafeQuery.mockReset();
+    mockSafeQuery.mockResolvedValue({ rows: [{ '?column?': 1 }], rowCount: 1 });
   });
 
   it('GET /health/live should return live status', async () => {
@@ -20,11 +31,29 @@ describe('Health Endpoints Verification', () => {
     expect(res.body.message).toBe('Server is alive');
   });
 
-  it('GET /health/ready should return 503 if dependencies are unknown', async () => {
+  it('GET /health/ready should return 200 when SELECT 1 succeeds even if Database registry is unknown', async () => {
+    const res = await request(app).get('/health/ready');
+    expect(res.status).toBe(200);
+    expect(res.body.status).toBe('healthy');
+    expect(mockSafeQuery).toHaveBeenCalled();
+  });
+
+  it('GET /health/ready should return 503 when SELECT 1 fails', async () => {
+    mockSafeQuery.mockResolvedValue(null);
+
     const res = await request(app).get('/health/ready');
     expect(res.status).toBe(503);
     expect(res.body.status).toBe('unhealthy');
-    expect(res.body.failedDependencies).toContain('Database');
+    expect(res.body.failedDependencies.some((d: { name: string }) => d.name === 'Database')).toBe(true);
+  });
+
+  it('GET /health/ready should return 503 when Database registry is unhealthy', async () => {
+    updateDependencyHealth('Database', 'unhealthy', 0, 'connection refused');
+
+    const res = await request(app).get('/health/ready');
+    expect(res.status).toBe(503);
+    expect(res.body.status).toBe('unhealthy');
+    expect(res.body.failedDependencies.some((d: { name: string }) => d.name === 'Database')).toBe(true);
   });
 
   it('GET /health/ready should return 200 when Database is healthy (optional deps may fail)', async () => {
@@ -47,6 +76,11 @@ describe('Health Endpoints Verification', () => {
   });
 
   it('GET /health/full should return comprehensive status', async () => {
+    updateDependencyHealth('Database', 'healthy', 10);
+    updateDependencyHealth('Redis', 'healthy', 5);
+    updateDependencyHealth('Neo4j', 'healthy', 5);
+    updateDependencyHealth('Pinecone', 'healthy', 5);
+
     const res = await request(app).get('/health/full');
     expect(res.status).toBe(200);
     expect(res.body.status).toBe('healthy');
